Move helpers to `cli_test_helpers.py`.
